+++ AsyncLibrary/__init__.py
@@ -1,1 +1,1 @@
-from .async import AsyncLibrary
+from .robot_async import AsyncLibrary
